Updated team grid
