+++ client/Components/SideBarContext/AboutContext.jsx
@@ -1,0 +1,5 @@
+import { React } from '../../../deps.ts';
+
+const AboutContext = (props) => <div></div>;
+
+export default AboutContext;
